ai-app/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pydantic-settings>=2.6.1
alembic>=1.14.0
langchain>=0.3.7
langchain-core>=0.3.18
langchain-groq>=0.2.0
langgraph>=0.2.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp-proto-grpc>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
opentelemetry-instrumentation-sqlalchemy>=0.50b0
pyyaml>=6.0
